+++ packages/ui/src/upload/avatar.tsx
@@ -47,32 +47,32 @@ export const UploadAvatar = ({
     >
       <input {...getInputProps()} />
       {preview && <img src={preview} alt="Preview" />}
-      <IconWrap hasPreview={!!preview}>
+      <PreviewOverlay isVisible={!preview}>
         {isDragActive ? (
           <DropIcon className="h-5 w-5" />
         ) : (
           <EditIcon className="h-5 w-5" />
         )}
-      </IconWrap>
+      </PreviewOverlay>
     </div>
   );
 };
 
-const IconWrap = ({
+const PreviewOverlay = ({
   children,
-  hasPreview,
+  isVisible,
 }: {
-  hasPreview: boolean;
+  isVisible: boolean;
   children: React.ReactElement;
 }) => {
   return (
     <div
       className={clsx(
         'absolute flex items-center justify-center',
-        'h-7 w-7 rounded transition-all duration-200',
+        'p-2 rounded transition-all duration-200',
         'group-hover:bg-background/80',
         'group-hover:opacity-100',
-        hasPreview ? 'opacity-0' : 'opacity-70 bg-background/20',
+        isVisible ? 'opacity-70 bg-background/20' : 'opacity-0',
       )}
     >
       {children}
