+++ packages/ui/src/upload/file.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
 
-import { FileUploader } from './upload-file';
+import { FileUploader } from './file';
 
 const meta = {
   component: FileUploader,
