+++ packages/ui/src/upload/image.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { ImageUploader } from './uploader-image';
+import { ImageUploader } from './image';
 import { within } from '@storybook/testing-library';
 import { expect } from '@storybook/jest';
